Refactor syncer unit tests to separate files
